[A64/Assembler] Fix source map index comparisons
- Cast source_map size to avoid signed/unsigned warnings

+++ src/xenia/cpu/backend/a64/a64_assembler.cc
@@ -130,11 +130,11 @@ void A64Assembler::DumpMachineCode(
     auto code_offset =
         uint32_t(code_ptr - reinterpret_cast<uint8_t*>(machine_code));
     if (code_offset >= next_code_offset &&
-        source_map_index < source_map.size()) {
+        source_map_index < static_cast<int>(source_map.size())) {
       auto& source_map_entry = source_map[source_map_index];
       str->AppendFormat("{:08X} ", source_map_entry.guest_address);
       ++source_map_index;
-      next_code_offset = source_map_index < source_map.size()
+      next_code_offset = source_map_index < static_cast<int>(source_map.size())
                              ? source_map[source_map_index].code_offset
                              : UINT_MAX;
     } else {
